% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correDimEst.R
\name{C2}
\alias{C2}
\title{Calculating proportion of the pairwise L2-distance that is less than epsilon}
\usage{
C2(epsilon, distance)
}
\arguments{
\item{epsilon}{Epsilon value.}

\item{distance}{Distance object.}
}
\value{
Proportion of the pairwise L2-distance that is less than epsilon.
}
\description{
Calculating proportion of the pairwise L2-distance that is less than epsilon
}

