% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correDimEst.R
\name{correDimEst}
\alias{correDimEst}
\title{Correlation dimension estimation}
\usage{
correDimEst(num, distance)
}
\arguments{
\item{num}{Number of epsilons.}

\item{distance}{Distance object (upper triangle) from dist() fn.}
}
\value{
a list of log(epsilons) (x), y and derivatives of y.
}
\description{
Takes in pair-wise distance of data and number of epsilons to estimate intrinsic dimension of data.
Function storing the logC2(epsi) values and the numerical derivative of logC2 w.r.t log(epsi).
}

