% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kplanes.R
\name{kplanes}
\alias{kplanes}
\title{Local principal component analysis (LPCA).
Extends k-means to k-planes.
In k-means centers are singlton points, but in k-planes centers are the
d-dimensional hyperplanes where d is the intrinsic dimension of the data.}
\usage{
kplanes(X, k, d, iter.max = 50, thresh = 0)
}
\arguments{
\item{X}{Data.}

\item{k}{Number of clusters.}

\item{d}{Intrinsic dimension.}

\item{iter.max}{Maximum number of iterations.}

\item{thresh}{Threshold for stopping, percentage change of normalized reconstruction error.}
}
\value{
a list of cluster membership (indicator), representaion (X.rep), average normalized
reconstruction error (error).
}
\description{
Takes in data, number of clusters, intrinsic dimension,
maximum iteration number and threshold for stopping
}
\examples{
############################## example I: Open box
## package for 3d plot
library(rgl)

## load data
data(OpenBox)

## intrinsic dimension of the data
trueDim = 2
## threshold for kplanes stopping (relative reconstruction error change ratio)
thresh = 0
## number of clusters
K = 6
set.seed(0)
temp = kplanes(OpenBox, k=K, d=trueDim, thresh=thresh)
OpenBox_rep = temp[[2]]
cluster_id = temp[[1]]
error_rep = temp[[3]]
open3d()
plot3d(OpenBox,col=cluster_id,xlim=c(0,2),ylim=c(0,2),zlim=c(0,2))


############################## example II: Swiss roll
## package for 3d plot
library(rgl)

## load data
data(SwissRoll)

## intrinsic dimension of the data
trueDim = 2

## threshold for kplanes stopping (relative reconstruction error change ratio)
thresh = 0

## number of clusters
K = 8

set.seed(0)
temp = kplanes(SwissRoll, k=K, d=trueDim, iter.max=1000, thresh=thresh)
SwissRoll_rep = temp[[2]]
cluster_id = temp[[1]]
error_rep = temp[[3]]

open3d()
plot3d(SwissRoll,col=cluster_id,xlim=c(0,2),ylim=c(0,2),zlim=c(0,2))

}

